requirements.txt
# Core Libraries
langchain
langchain-community
langchain-huggingface
langchain-google-genai
streamlit
python-dotenv

# Data & Vector Stores
chromadb
sentence-transformers
faiss-cpu

# Web & APIs
requests
beautifulsoup4
semanticscholar

# Database
psycopg2-binary